requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.28
Jinja2==3.1.3
python-multipart==0.0.9
Pillow==10.3.0
nanoid==2.0.0

itsdangerous>=2.1
